(long) in -sizeof() to fix overflow in conversion from 'long long unsigned int' to 'long int' changes value from '18446744073709551608' to '-8' [-Woverflow]

+++ lm/search_trie.cc
@@ -88,7 +88,7 @@ class BackoffMessages {
         }
         if (!HasExtension(weights.backoff)) {
           weights.backoff = kExtensionBackoff;
-          UTIL_THROW_IF(fseek(unigrams, -sizeof(weights), SEEK_CUR), util::ErrnoException, "Seeking backwards to denote unigram extension failed.");
+          UTIL_THROW_IF(fseek(unigrams, -(long)sizeof(weights), SEEK_CUR), util::ErrnoException, "Seeking backwards to denote unigram extension failed.");
           util::WriteOrThrow(unigrams, &weights, sizeof(weights));
         }
         const ProbPointer &write_to = *reinterpret_cast<const ProbPointer*>(current_ + sizeof(WordIndex));
